리팩터링: 사다리 생성기 삭제. 사다리 생성 로직은 사다리 객체가 가지도록 변경
