+++ src/main/java/ladder/controller/LadderGameController.java
@@ -1,6 +1,6 @@
 package ladder.controller;
 
-import ladder.domain.generator.Generator;
+import ladder.domain.generator.RandomLadderLineGenerator;
 import ladder.view.LadderRenderer;
 import ladder.view.NameRenderer;
 import ladder.view.Prompt;
@@ -8,7 +8,6 @@ import ladder.view.MessageRenderer;
 import ladder.domain.Ladder;
 import ladder.domain.Name;
 import ladder.domain.Names;
-import ladder.domain.generator.RandomLadderGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -52,9 +51,7 @@ public class LadderGameController {
     }
 
     private void ladderGenerationPhase() {
-        this.ladder = Ladder.of(
-                new RandomLadderGenerator(depth, this.names.size())
-        );
+        this.ladder = Ladder.of(new RandomLadderLineGenerator(this.names.size()), depth);
     }
 
     private void printPhase() {
